requirements.txt
flask
flask_sqlalchemy
flask_cors
requests-html
nltk
psycopg2
